pyproject.toml
[project]
name = "kaggle-mcp"
version = "0.1.0"
description = "MCP Server for interacting with Kaggle API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kaggle",
    "python-dotenv",
    "mcp-server",
    # "uvicorn[standard]", # Removed dependency
]
scripts = { kaggle-mcp = "server:run_server" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Correct structure for automatic package discovery
[tool.setuptools.packages.find]
where = ["src"]
# Exclude the 'datasets' directory if found within 'src'
exclude = ["datasets*"]
